move function for save and load data from files into 'utils'
