+++ utils.py
@@ -1,3 +1,4 @@
+import json
 import re
 from collections import namedtuple
 
@@ -17,3 +18,18 @@ def get_repos_info(links, pattern):
                     repo_name = result.group("name").lower()
                     repos_info.add(RepoMetaData(owner=owner, repo_name=repo_name))
     return sorted(repos_info)
+
+
+def save_json_data_and_return(data, path):
+    """
+    Saves data as a json file and returns the data.
+    """
+    path.write_text(json.dumps(data, indent=4))
+    return data
+
+
+def load_json_data(path):
+    """
+    Loads data from a json file and returns it.
+    """
+    return json.loads(path.read_text(encoding="UTF-8")) if path.is_file() else {}
